Arend Proof Search: Decouple matching from structural search to reuse it in "search everywhere"
